When mining and refining resources, Megajoules should only be consumed when there are input materials available AND there's storage onboard for the output materials.
This patch prevents a deadlock refueling condition. For example, on mun, a 1.25m generator + sethanes fission reactor cannot supply enough power to both mine alumina and electrolyse it into aluminum and oxidizer. If both options are enabled on the ISRU, the mining will take precidence and starve the electrolysis process of megajoules even when the alumina storage on the ship is full. With this patch applied, the mining function will stop consuming megajoules when there is no more room to store the alumina, so there will be power for the electrolysis to consume the alumina. This allows the ship to be refueled without requiring the player to constantly turn on and off the alumina mining option (although it will still be slower than if sufficient power generation were available).